So I figured out what was up with REST implementing it within our current malware wasnt all that hard well at least editing the code that is, truth be told the hard part was just figuring out what REST is and how Im supposed to have the AWS server and our little python thingy talk to each other this way. I.. succeded quite cleanly, I can technically satisfy the lab, but it is far from being true malware.
anyway, malware2.py is my latest in computer japery, the final
pruduct is supposed to be in two executables, one being the
malware, doing only encryption. and the other being an unlocker.
malware2.py is a prototype right now with both of those functions
back to back in the program.
also I have pushed a db.json file to the git, dont worry about
that, its a file I mean to transfer over to the AWS server.
Server is going to use that file to hold the key pairs.